refactor: 스토리북 예제 복사 함수 모듈화
